+++ package.json
@@ -10,7 +10,7 @@
     "build": "./node_modules/.bin/tsc",
     "prepack": "rm -rf ./lib && export NODE_ENV=production && ./node_modules/.bin/tsc --project tsconfig.production.json",
     "postversion": "git push origin $(git rev-parse --abbrev-ref HEAD) --tags",
-    "test": "./node_modules/.bin/mocha src/**/*.spec.ts",
+    "test": "vitest --run",
     "ci:publish": "changeset publish && git push --follow-tags",
     "ci:version": "changeset version && pnpm install --no-frozen-lockfile"
   },
@@ -19,16 +19,13 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.26.2",
-    "@types/chai": "^4.0.10",
-    "@types/mocha": "^2.2.44",
     "@types/node": "^22.15.1",
     "lerna": "^8.2.3",
     "aws-sdk": "^2.168.0",
-    "chai": "^4.1.2",
-    "mocha": "^4.0.1",
     "prettier": "^3.3.3",
     "ts-node": "^10.9.2",
-    "typescript": "^4.9.5"
+    "typescript": "^4.9.5",
+    "vitest": "^1.6.0"
   },
   "packageManager": "pnpm@10.12.4+sha512.5ea8b0deed94ed68691c9bad4c955492705c5eeb8a87ef86bc62c74a26b037b08ff9570f108b2e4dbd1dd1a9186fea925e527f141c648e85af45631074680184",
   "dependencies": {
