--- test/mocha.opts
@@ -1,2 +1,0 @@
---require ts-node/register
---watch-extensions ts
