+++ vitest.config.unit.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    include: ['./tests/**/*.test.ts'],
+    reporters: 'verbose',
+    environment: 'node'
+  }
+});
